+++ docs/conf.py
@@ -29,6 +29,7 @@ source_suffix = {
 }
 
 master_doc = "index"
+templates_path = ["_templates"]
 exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 
 autodoc_typehints = "description"